Move to separate files each intro content type composable
