+++ feature/trip-planner/ui/src/commonMain/kotlin/xyz/ksharma/krail/trip/planner/ui/intro/IntroContent.kt
@@ -74,159 +74,6 @@ val trip = Trip(
     toStopId = "2"
 )
 
-@Composable
-fun IntroContentSaveTrips(
-    tagline: String,
-    modifier: Modifier = Modifier,
-    style: String, // hexCode - // todo - see if it can be color instead.
-) {
-    Column(
-        modifier = modifier
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-            OriginDestination(
-                trip = trip,
-                timeLineColor = KrailTheme.colors.onSurface,
-            )
-
-            Divider()
-        }
-
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            // TODO if tapped more than 5 times, show confetti of stars.
-            var isTripSaved by rememberSaveable { mutableStateOf(false) }
-            ActionButton(
-                onClick = { isTripSaved = !isTripSaved },
-                contentDescription = if (isTripSaved) {
-                    "Remove Saved Trip"
-                } else {
-                    "Save Trip"
-                },
-                modifier = Modifier.size(80.dp),
-            ) {
-                Image(
-                    painter = if (isTripSaved) {
-                        painterResource(Res.drawable.ic_star_filled)
-                    } else {
-                        painterResource(Res.drawable.ic_star)
-                    },
-                    contentDescription = null,
-                    colorFilter = ColorFilter.tint(style.hexToComposeColor()),
-                    modifier = Modifier.size(56.dp),
-                )
-            }
-        }
-
-        TagLineWithEmoji(
-            tagline = tagline,
-            emoji = "\uD83C\uDF1F",
-            tagColor = style.hexToComposeColor()
-        )
-    }
-}
-
-@Composable
-fun IntroContentRealTime(
-    tagline: String,
-    style: String, // hexCode - // todo - see if it can be color instead.
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Column(verticalArrangement = Arrangement.spacedBy(10.dp)) {
-            LegView(
-                routeText = "Manly to Circular Quay",
-                transportModeLine = TransportModeLine(
-                    transportMode = TransportMode.Ferry(),
-                    lineName = "MFF",
-                ),
-                stops = stopsList(),
-            )
-        }
-
-        TagLineWithEmoji(
-            tagline = tagline, emoji = "\uD83D\uDE80",
-            tagColor = style.hexToComposeColor()
-        )
-    }
-}
-
-@Composable
-fun IntroContentAlerts(
-    tagline: String,
-    style: String, // hexCode - // todo - see if it can be color instead.
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Column(
-            verticalArrangement = Arrangement.spacedBy(12.dp),
-        ) {
-
-            var displayAlert by rememberSaveable { mutableStateOf(false) }
-
-            AlertButton(
-                dimensions = ButtonDefaults.smallButtonSize(),
-                onClick = { displayAlert = !displayAlert },
-            ) { Text(text = "2 Alerts") }
-
-            AnimatedAlerts(displayAlert)
-        }
-
-        TagLineWithEmoji(
-            tagline = tagline, emoji = "⚠", emojiColor = style.hexToComposeColor(),
-            tagColor = style.hexToComposeColor()
-        )
-    }
-}
-
-@Composable
-fun AnimatedAlerts(displayAlert: Boolean) {
-    AnimatedVisibility(
-        visible = displayAlert,
-        enter = fadeIn(animationSpec = tween(300)) +
-                slideInVertically(
-                    initialOffsetY = { -20 },
-                    animationSpec = tween(300),
-                ),
-        exit = fadeOut(animationSpec = tween(300)) +
-                slideOutVertically(
-                    targetOffsetY = { -20 },
-                    animationSpec = tween(300),
-                )
-    ) {
-        Column(verticalArrangement = Arrangement.spacedBy(8.dp)) {
-            CollapsibleAlert(
-                serviceAlert = ServiceAlert(
-                    heading = "Running late. Please allow extra travel time.", message = "",
-                ),
-                index = 1,
-                collapsed = false,
-                onClick = {},
-            )
-
-            CollapsibleAlert(
-                serviceAlert = ServiceAlert(
-                    heading = "Platforms may change, listen for announcements.", message = "",
-                ),
-                index = 2,
-                collapsed = false,
-                onClick = {},
-            )
-        }
-    }
-}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -279,98 +126,7 @@ fun IntroContentPlanTrip(
 }
 
 @Composable
-fun IntroContentInviteFriends(
-    tagline: String,
-    style: String, // hexCode - // todo - see if it can be color instead.
-    onShareClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            Text(
-                text = "TRAVELLING SOLO?\nNAH THANKS",
-                style = KrailTheme.typography.introTagline,
-                color = style.hexToComposeColor(),
-            )
-        }
-
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(64.dp)
-                    .clip(CircleShape)
-                    .background(color = style.hexToComposeColor())
-                    .clickable(
-                        indication = null,
-                        role = Role.Button,
-                        interactionSource = remember { MutableInteractionSource() },
-                        onClick = onShareClick
-                    ),
-                contentAlignment = Alignment.Center,
-            ) { // TODO - show diff. image for ios / android
-                Image(
-                    painter = painterResource(Res.drawable.ic_ios_share),
-                    contentDescription = null,
-                    colorFilter = ColorFilter.tint(Color.White),
-                    modifier = Modifier.size(48.dp),
-                )
-            }
-        }
-
-        TagLineWithEmoji(
-            tagline = tagline, emoji = "\uD83D\uDC95",
-            tagColor = style.hexToComposeColor()
-        )
-    }
-}
-
-@Composable
-fun IntroContentSelectTransportMode(
-    tagline: String,
-    style: String, // hexCode - // todo - see if it can be color instead.
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = modifier
-            .verticalScroll(rememberScrollState()),
-        verticalArrangement = Arrangement.SpaceBetween,
-    ) {
-        Column(verticalArrangement = Arrangement.spacedBy(12.dp)) {
-            var selectedModes by rememberSaveable { mutableStateOf(setOf<TransportMode>()) }
-
-            TransportMode.values().forEach { mode ->
-                TransportModeChip(
-                    transportMode = mode,
-                    selected = selectedModes.contains(mode),
-                    onClick = {
-                        selectedModes = if (selectedModes.contains(mode)) {
-                            selectedModes - mode
-                        } else selectedModes + mode
-                    },
-                    //modifier = Modifier.padding(horizontal = 10.dp),
-                )
-            }
-        }
-
-        TagLineWithEmoji(
-            tagline = tagline,
-            emoji = "\uD83D\uDE0E",
-            tagColor = style.hexToComposeColor()
-        )
-    }
-}
-
-
-@Composable
-private fun TagLineWithEmoji(
+internal  fun TagLineWithEmoji(
     tagline: String,
     emoji: String,
     emojiColor: Color? = null,
@@ -394,16 +150,3 @@ private fun TagLineWithEmoji(
         )
     }
 }
-
-private fun stopsList() = persistentListOf(
-    TimeTableState.JourneyCardInfo.Stop(
-        name = "Manly, Wharf 2",
-        time = "10:10 AM",
-        isWheelchairAccessible = true,
-    ),
-    TimeTableState.JourneyCardInfo.Stop(
-        name = "Circular Quay, Wharf 2",
-        time = "10:30 AM",
-        isWheelchairAccessible = true,
-    ),
-)
